@@ -3,18 +3,26 @@ apply plugin: 'AndResGuard'
 apply plugin: 'org.greenrobot.greendao'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.massky.sraum"
         minSdkVersion 15
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         //添加如下配置就OK了
         javaCompileOptions { annotationProcessorOptions { includeCompileClasspath = true } }
-        ndk { abiFilters "armeabi"}
-//        multiDexEnabled true//设置为支持多dex输出模式
+        ndk {
+            //设置支持的SO库架构
+            abiFilters 'armeabi', 'armeabi-v7a', 'x86'//, , 'x86_64', 'arm64-v8a'方便去除模擬器的，縮小apk體積
+        }
+
+        javaCompileOptions { annotationProcessorOptions { includeCompileClasspath = true } }
+//        multiDexEnabled true//设置
+        // 为支持多dex输出模式
+//        multiDexEnabled true
     }
 
     buildTypes {
