@@ -4,7 +4,7 @@ import android.app.Dialog;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.telephony.TelephonyManager;
 import android.view.LayoutInflater;
 import android.view.View;
