@@ -1,8 +1,8 @@
 package com.ipcamera.demo.adapter;
 
 import android.os.Parcelable;
-import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.PagerAdapter;
+import androidx.viewpager.widget.ViewPager;
 import android.view.View;
 
 import java.util.List;
