@@ -6,19 +6,16 @@ import android.content.Context;
 import android.content.Intent;
 import android.media.MediaPlayer;
 import android.os.Bundle;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.text.TextUtils;
 import android.util.Log;
-import com.AddTogenInterface.AddTogglenInterfacer;
+
 import com.alibaba.fastjson.JSON;
 import com.massky.sraum.User;
 import com.massky.sraum.Util.LogUtil;
 import com.massky.sraum.Util.MusicUtil;
-import com.massky.sraum.Util.MyOkHttp;
-import com.massky.sraum.Util.Mycallback;
 import com.massky.sraum.Util.SharedPreferencesUtil;
 import com.massky.sraum.Util.TokenUtil;
-import com.massky.sraum.Utils.ApiHelper;
 import com.massky.sraum.activity.FastEditPanelActivity;
 import com.massky.sraum.activity.MainGateWayActivity;
 import com.massky.sraum.base.BaseActivity;
@@ -27,7 +24,7 @@ import com.massky.sraum.fragment.HomeFragment;
 import com.massky.sraum.fragment.SceneFragment;
 import org.json.JSONException;
 import org.json.JSONObject;
-import java.util.ArrayList;
+
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
