@@ -2,9 +2,8 @@ package com.massky.sraum.activity;
 
 import android.content.Intent;
 import android.os.Handler;
-import android.os.Looper;
 import android.os.Message;
-import android.support.percent.PercentRelativeLayout;
+import androidx.percentlayout.widget.PercentRelativeLayout;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
