@@ -3,7 +3,7 @@ package com.massky.sraum.service;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import com.google.gson.Gson;
 import com.massky.sraum.Util.ICallback;
 import com.massky.sraum.Util.IConnectTcpback;
