@@ -2,10 +2,9 @@ package com.pullableview;
 
 import android.content.Context;
 import android.util.AttributeSet;
-import android.widget.ImageView;
 
 
-public class PullableImageView extends android.support.v7.widget.AppCompatImageView implements Pullable
+public class PullableImageView extends androidx.appcompat.widget.AppCompatImageView implements Pullable
 {
 
 	public PullableImageView(Context context)
