@@ -3,11 +3,14 @@ package com.massky.sraum.activity;
 import android.app.FragmentManager;
 import android.app.FragmentTransaction;
 import android.content.Intent;
+import android.os.Handler;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ImageView;
 
 import com.AddTogenInterface.AddTogglenInterfacer;
+import com.gizwits.gizwifisdk.api.GizWifiDevice;
+import com.gizwits.gizwifisdk.enumration.GizWifiErrorCode;
 import com.massky.sraum.R;
 import com.massky.sraum.User;
 import com.massky.sraum.Util.DialogUtil;
@@ -25,9 +28,15 @@ import com.massky.sraum.fragment.GatewayDialogFragment;
 import com.massky.sraum.widget.ListViewForScrollView;
 import com.yanzhenjie.statusview.StatusUtils;
 import com.yanzhenjie.statusview.StatusView;
+import com.yaokan.sdk.utils.Logger;
+import com.yaokan.sdk.utils.Utility;
+import com.yaokan.sdk.wifi.DeviceManager;
+import com.yaokan.sdk.wifi.GizWifiCallBack;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 
@@ -71,19 +80,19 @@ public class SelectZigbeeDeviceActivity extends BaseActivity {
     //11-紧急按钮，12-久坐报警器，13-烟雾报警器，14-天然气报警器，15-智能门锁，16-直流电阀机械手
 
     //wifi类型
+    //wifi类型
     private String[] types_wifi = { //红外转发器类型暂定为hongwai,遥控器类型暂定为yaokong
             "hongwai", "yaokong", "101", "103"
     };
 
     private int[] icon_wifi = {
-            R.drawable.icon_maoyan,
-            R.drawable.icon_maoyan,
-            R.drawable.icon_maoyan,
-            // R.drawabicon_maoyan,
-            R.drawable.icon_maoyan,
+            R.drawable.hongwai_s,
+            R.drawable.icon_type_yaokongqi,
+            R.drawable.icon_type_shexiangtou_90,
+            // R.drawable.icon_type_pmmofang_90,
+            R.drawable.icon_keshimenling
 
     };
-
     private int[] iconNam_wifi = {R.string.hongwai, R.string.yaokongqi, R.string.shexiangtou, R.string.keshimenling};//, R.string.pm_mofang
     private int[] iconName = {R.string.yijianlight, R.string.liangjianlight, R.string.sanjianlight, R.string.sijianlight,
             R.string.yilutiaoguang1, R.string.lianglutiaoguang1, R.string.sanlutiao1, R.string.window_panel1
@@ -96,6 +105,22 @@ public class SelectZigbeeDeviceActivity extends BaseActivity {
     private GatewayDialogFragment newGatewayFragment;
     private DialogUtil dialogUtil;
 
+    private List<Map> list_hand_scene = new ArrayList<>();
+
+    /**
+     * 小苹果绑定列表
+     */
+    private DeviceManager mDeviceManager;
+    List<GizWifiDevice> wifiDevices = new ArrayList<GizWifiDevice>();
+    private List<String> deviceNames = new ArrayList<String>();
+    private GizWifiDevice mGizWifiDevice = null;
+    private List<Map> wifi_apple_list = new ArrayList<>();
+
+    private String TAG = "robin debug";
+    private String mac;
+    private String number;
+    private Handler handler = new Handler();
+
     @Override
     protected int viewId() {
         return R.layout.add_device_act;
@@ -107,6 +132,12 @@ public class SelectZigbeeDeviceActivity extends BaseActivity {
         dialogUtil = new DialogUtil(this);
         initDialog();
         initGatewayDialog();
+        onWifi();
+    }
+
+    private void onWifi() {
+        mDeviceManager = DeviceManager
+                .instanceDeviceManager(getApplicationContext());
     }
 
     @Override
@@ -294,6 +325,30 @@ public class SelectZigbeeDeviceActivity extends BaseActivity {
         });
 
 
+        mac_wifi_dev_id.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                Intent intent_wifi = null;
+                switch (types_wifi[position]) {
+                    case "101":
+                    case "103":
+                        intent_wifi = new Intent(SelectZigbeeDeviceActivity.this, AddWifiDevActivity.class);
+                        intent_wifi.putExtra("type", types_wifi[position]);
+                        startActivity(intent_wifi);
+                        break;
+                    case "102":
+                        break;
+                    case "hongwai":
+                        intent_wifi = new Intent(SelectZigbeeDeviceActivity.this, AddWifiDevActivity.class);
+                        intent_wifi.putExtra("type", types_wifi[position]);
+                        startActivity(intent_wifi);
+                        break;
+                    case "yaokong":
+                        getOtherDevices();
+                        break;
+                }
+            }
+        });
     }
 
     /**
@@ -316,6 +371,204 @@ public class SelectZigbeeDeviceActivity extends BaseActivity {
         sraum_getAllGateWays(map);
     }
 
+    /**
+     * 获取门磁等第三方设备
+     */
+    private void getOtherDevices() {
+        if (dialogUtil != null) {
+            dialogUtil.loadDialog();
+        }
+        Map<String, String> mapdevice = new HashMap<>();
+        mapdevice.put("token", TokenUtil.getToken(SelectZigbeeDeviceActivity.this));
+//        mapdevice.put("boxNumber", TokenUtil.getBoxnumber(SelectSensorActivity.this));
+        MyOkHttp.postMapString(ApiHelper.sraum_getWifiAppleInfos
+                , mapdevice, new Mycallback(new AddTogglenInterfacer() {
+                    @Override
+                    public void addTogglenInterfacer() {//刷新togglen数据
+                        getOtherDevices();
+                    }
+                }, SelectZigbeeDeviceActivity.this, dialogUtil) {
+                    @Override
+                    public void onError(Call call, Exception e, int id) {
+                        super.onError(call, e, id);
+                    }
+
+                    @Override
+                    public void pullDataError() {
+                        super.pullDataError();
+                    }
+
+                    @Override
+                    public void emptyResult() {
+                        super.emptyResult();
+                    }
+
+                    @Override
+                    public void wrongToken() {
+                        super.wrongToken();
+                        //重新去获取togglen,这里是因为没有拉到数据所以需要重新获取togglen
+                    }
+
+                    @Override
+                    public void wrongBoxnumber() {
+                        super.wrongBoxnumber();
+                    }
+
+                    @Override
+                    public void onSuccess(final User user) {
+                        list_hand_scene = new ArrayList<>();
+                        for (int i = 0; i < user.controllerList.size(); i++) {
+                            Map<String, String> mapdevice = new HashMap<>();
+                            mapdevice.put("name", user.controllerList.get(i).name);
+                            mapdevice.put("number", user.controllerList.get(i).number);
+                            mapdevice.put("type", user.controllerList.get(i).type);
+                            mapdevice.put("controllerId", user.controllerList.get(i).controllerId);
+                            list_hand_scene.add(mapdevice);
+                        }
+
+                        if (list_hand_scene.size() == 1) {
+                            if (wifiDevices.size() != 0) {
+                                choose_the_brand(0);
+                            } else {
+//                                update(mDeviceManager.getCanUseGizWifiDevice());
+                                ToastUtil.showToast(SelectZigbeeDeviceActivity.this, "请与" + list_hand_scene.get(0)
+                                        .get("name").toString()
+                                        +
+                                        "在同一网络后再添加");
+                            }
+                        } else {
+                            Intent intent_wifi =
+                                    new Intent(SelectZigbeeDeviceActivity.this,
+                                            SelectInfraredForwardActivity.class);
+                            intent_wifi.putExtra("list_hand_scene", (Serializable) list_hand_scene);
+                            startActivity(intent_wifi);
+                        }
+                    }
+                });
+    }
+
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        mDeviceManager.setGizWifiCallBack(mGizWifiCallBack);
+        update(mDeviceManager.getCanUseGizWifiDevice());
+    }
+
+
+    private GizWifiCallBack mGizWifiCallBack = new GizWifiCallBack() {
+
+        @Override
+        public void didUnbindDeviceCd(GizWifiErrorCode result, String did) {
+            super.didUnbindDeviceCd(result, did);
+            if (result == GizWifiErrorCode.GIZ_SDK_SUCCESS) {
+                // 解绑成功
+                Logger.d(TAG, "解除绑定成功");
+            } else {
+                // 解绑失败
+                Logger.d(TAG, "解除绑定失败");
+            }
+        }
+
+        @Override
+        public void didBindDeviceCd(GizWifiErrorCode result, String did) {    //
+            super.didBindDeviceCd(result, did);
+            if (result == GizWifiErrorCode.GIZ_SDK_SUCCESS) {//
+                // 绑定成功
+                Logger.d(TAG, "绑定成功");
+            } else {
+                // 绑定失败
+                Logger.d(TAG, "绑定失败");
+            }
+        }
+
+        @Override
+        public void didSetSubscribeCd(GizWifiErrorCode result, GizWifiDevice device, boolean isSubscribed) {
+            super.didSetSubscribeCd(result, device, isSubscribed);
+            if (result == GizWifiErrorCode.GIZ_SDK_SUCCESS) {
+                // 解绑成功
+                Logger.d(TAG, (isSubscribed ? "订阅" : "取消订阅") + "成功");
+            } else {
+                // 解绑失败
+                Logger.d(TAG, "订阅失败");
+            }
+        }
+
+        @Override
+        public void discoveredrCb(GizWifiErrorCode result,
+                                  List<GizWifiDevice> deviceList) {
+            Logger.d(TAG,
+                    "discoveredrCb -> deviceList size:" + deviceList.size()
+                            + "  result:" + result);
+            switch (result) {
+                case GIZ_SDK_SUCCESS:
+                    Logger.e(TAG, "load device  sucess");
+                    update(deviceList);
+//                    if(deviceList.get(0).getNetStatus()==GizWifiDeviceNetStatus.GizDeviceOffline)
+
+                    break;
+                default:
+                    break;
+            }
+        }
+    };
+
+    /**
+     * 跳转到选择品牌界面
+     *
+     * @param position
+     */
+    private void choose_the_brand(int position) {
+        mac = (String) list_hand_scene.get(position).get("controllerId");
+        number = list_hand_scene.get(position).get("number").toString();
+        //去根据mac去服务器端下载GizWifiDevice
+        String apple_name = "";
+        for (int i = 0; i < list_hand_scene.size(); i++) {
+            if (list_hand_scene.get(i).get("controllerId").equals(mac)) {
+                apple_name = list_hand_scene.get(i).get("name").toString();
+
+            }
+        }
+        get_to_wifi(mac, apple_name);
+    }
+
+    private void get_to_wifi(String mac, String apple_name) {
+
+        for (int i = 0; i < wifiDevices.size(); i++) {
+            if (wifiDevices.get(i).getMacAddress().equals(mac)) {
+                mGizWifiDevice = wifiDevices.get(i);
+            }
+        }
+        if (!Utility.isEmpty(mGizWifiDevice)) { //
+            mDeviceManager.bindRemoteDevice(mGizWifiDevice);
+            final GizWifiDevice finalMGizWifiDevice = mGizWifiDevice;
+            handler.postDelayed(new Runnable() {
+                @Override
+                public void run() {
+                    mDeviceManager.setSubscribe(finalMGizWifiDevice, true);
+                }
+            }, 1000);
+        } else {
+            ToastUtil.showToast(SelectZigbeeDeviceActivity.this, "请与" + apple_name
+                    +
+                    "在同一网络后再控制");
+            return;
+        }
+        toControlApplianAct();
+    }
+
+    private void toControlApplianAct() {
+        Intent intent = new Intent(SelectZigbeeDeviceActivity.this, SelectControlApplianceActivity.class);
+        if (mGizWifiDevice == null) {
+            return;
+        }
+        intent.putExtra("GizWifiDevice", mGizWifiDevice);
+//        intent.putExtra("tid", getIntent().getSerializableExtra("tid"));
+        intent.putExtra("number", number);
+        startActivity(intent);
+    }
+
+
     @Override
     public void onClick(View v) {
         switch (v.getId()) {
@@ -323,5 +576,44 @@ public class SelectZigbeeDeviceActivity extends BaseActivity {
                 SelectZigbeeDeviceActivity.this.finish();
                 break;
         }
+    }
+
+    void update(List<GizWifiDevice> gizWifiDevices) {
+        GizWifiDevice mGizWifiDevice = null;
+
+
+        if (gizWifiDevices == null) {
+            deviceNames.clear();
+        } else if (gizWifiDevices != null && gizWifiDevices.size() >= 1) {
+//            wifiDevices.clear();
+            wifiDevices.addAll(gizWifiDevices);
+            HashSet<GizWifiDevice> h = new HashSet<GizWifiDevice>(wifiDevices);
+            wifiDevices.clear();
+            for (GizWifiDevice gizWifiDevice : h) {
+                wifiDevices.add(gizWifiDevice);
+            }
+            deviceNames.clear();
+//            for (int i = 0; i < wifiDevices.size(); i++) {
+////                deviceNames.add(wifiDevices.get(i).getProductName() + "("
+////                        + wifiDevices.get(i).getMacAddress() + ") "
+////                        + getBindInfo(wifiDevices.get(i).isBind()) + "\n"
+////                        + getLANInfo(wifiDevices.get(i).isLAN()) + "  " + getOnlineInfo(wifiDevices.get(i).getNetStatus()));
+//                mGizWifiDevice = wifiDevices.get(i);
+//                // list_hand_scene
+//                // 绑定选中项
+//                if (!Utility.isEmpty(mGizWifiDevice)) { //
+//                    mDeviceManager.bindRemoteDevice(mGizWifiDevice);
+//                    final GizWifiDevice finalMGizWifiDevice = mGizWifiDevice;
+//                    handler.postDelayed(new Runnable() {
+//                        @Override
+//                        public void run() {
+//                            mDeviceManager.setSubscribe(finalMGizWifiDevice, true);
+//                        }
+//                    }, 1000);
+//                }
+//            }
+            //去绑定和订阅
+        }
+//        adapter.notifyDataSetChanged();
     }
 }
